refactor(company)!: rename bs to buzz
