+++ src/locales/el/company/buzz_adjective.ts
