+++ src/locales/el/company/buzz_noun.ts
