+++ src/locales/el/company/buzz_verb.ts
