+++ src/locales/en/company/buzz_adjective.ts
