+++ src/locales/en/company/buzz_noun.ts
