+++ src/locales/en/company/buzz_verb.ts
