+++ src/locales/es_MX/company/buzz_adjective.ts
