+++ src/locales/es_MX/company/buzz_noun.ts
