+++ src/locales/es_MX/company/buzz_verb.ts
