+++ src/locales/fa/company/buzz_adjective.ts
