+++ src/locales/fa/company/buzz_noun.ts
