+++ src/locales/fa/company/buzz_verb.ts
