+++ src/locales/fr/company/buzz_adjective.ts
