+++ src/locales/fr/company/buzz_noun.ts
