+++ src/locales/fr/company/buzz_verb.ts
