+++ src/locales/it/company/buzz_adjective.ts
