+++ src/locales/pl/company/buzz_adjective.ts
