+++ src/locales/pl/company/buzz_noun.ts
