+++ src/locales/pl/company/buzz_verb.ts
